address unsafe-eval CSP issues, additional cleanup (#2752)
