+++ libraries/botframework-streaming/src/index.ts
@@ -24,7 +24,6 @@ export { StreamingRequest } from './streamingRequest';
 export { StreamingResponse } from './streamingResponse';
 export { SubscribableStream } from './subscribableStream';
 export {
-    BrowserWebSocket,
     NodeWebSocket,
     NodeWebSocketFactory,
     NodeWebSocketFactoryBase,
