+++ libraries/botframework-streaming/src/utilities/index.ts
@@ -6,6 +6,4 @@
  * Licensed under the MIT License.
  */
 
-export * from './doesGlobalFileReaderExist';
-export * from './doesGlobalWebSocketExist';
 export * from './protocol-base';
