+++ libraries/botframework-streaming/tsconfig.json
@@ -11,7 +11,7 @@
     "types" : ["node"]
   },
   "exclude": [
-    "**/index-browser.ts",
+    "**/*browser*.ts",
     "browser",
     "lib",
     "node_modules"
